# public/javascripts/swagger.json
{"components":{"schemas":{"Image":{"properties":{"id":{"type":"integer"},"authorId":{"type":"integer"},"tags":{"type":"array","items":{"type":"string"}},"title":{"type":"string"},"likes":{"type":"integer"},"editorIds":{"type":"array","items":{"type":"integer"}},"folderId":{"type":"integer"},"imagePath":{"type":"string"}}},"Folder":{"properties":{"id":{"type":"integer"},"authorId":{"type":"integer"},"title":{"type":"string"}}}}},"servers":[{"url":"http://localhost:9000/v1","description":"Localhost only (for now!)"}],"openapi":"3.0.0","paths":{"/api/v1/images/{id}/file":{"get":{"operationId":"getImageFileById","parameters":[{"in":"path","name":"id","type":"integer","format":"int64","required":true}],"tags":["routes"]}},"/assets/{file}":{"get":{"operationId":"versioned","parameters":[{"in":"path","name":"file","type":"asset","required":true}],"tags":["routes"]}},"/api/v1/folders/{id}":{"get":{"responses":{"200":{"schema":{"$ref":"#/definitions/models.Folder"}},"404":{"description":"Folder with specified ID was not found"}},"operationId":"getById","parameters":[{"in":"path","name":"id","type":"integer","format":"int64","required":true}],"tags":["routes"]},"delete":{"responses":{"204":{"description":"Folder with specified id has been deleted"},"404":{"description":"Folder with specified ID was not found"}},"operationId":"delete","parameters":[{"in":"path","name":"id","type":"integer","format":"int64","required":true}],"tags":["routes"]}},"/api/v1/comments/{id}/likeCount":{"patch":{"operationId":"updateLikeCount","parameters":[{"in":"path","name":"id","type":"integer","format":"int64","required":true}],"tags":["routes"]}},"/api/v1/folders/author/{authorId}":{"get":{"responses":{"200":{"schema":{"$ref":"#/definitions/models.Folder"}},"404":{"description":"Folder with specified author ID was not found"}},"operationId":"getByAuthorId","parameters":[{"in":"path","name":"authorId","type":"integer","format":"int64","required":true}],"tags":["routes"]}},"/api/v1/comments/{id}/content":{"patch":{"operationId":"updateContent","parameters":[{"in":"path","name":"id","type":"integer","format":"int64","required":true}],"tags":["routes"]}},"/folders":{"get":{"summary":"Returns a list of all folders in the database.","responses":{"200":{"description":"A JSON array of folder objects.","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Image"}}}}}}},"post":{"summary":"Creates and returns a new image.","description":"Creates a new image with the given author ID, tags, title and image file. Automatically generates the ID, sets the likes to 0, assigns the author ID to the list of editor IDs and the folder ID to -1.","requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"authorId":{"type":"string"},"tags":{"type":"array","items":{"type":"string"}},"title":{"type":"string"},"image":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"A JSON object of the image.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Image"}}}}}}},"/api/v1/comments":{"post":{"requestBody":{"content":{"application/json":{"schema":{"type":"object","$ref":"#/definitions/dtos.NewComment"}}}},"responses":{"200":{"schema":{"$ref":"#/definitions/models.Comment"}}},"operationId":"create","tags":["routes"]},"get":{"responses":{"200":{"schema":{"$ref":"#/definitions/models.Comment"}}},"operationId":"getAll","tags":["routes"]}},"/images/{id}/likeCount":{"patch":{"summary":"Updates the number of likes of the image with the given ID.","parameters":[{"in":"path","name":"id","schema":{"type":"integer","minimum":1},"description":"The image ID."},{"in":"query","name":"likeCount","schema":{"type":"integer","description":"The new number of likes of the image."}}],"responses":{"200":{"description":"Newly updated number of likes.","content":{"application/json":{"schema":{"type":"integer"}}}},"404":{"description":"Image with specified ID was not found."}}}},"/images/{id}/editorIds":{"patch":{"summary":"Updates the editor id list of the image with the given ID.","parameters":[{"in":"path","name":"id","schema":{"type":"integer","minimum":1},"description":"The image ID."},{"in":"query","name":"editorIds","schema":{"type":"array","items":{"type":"integer"},"description":"The new list of editor IDs (longs)."}}],"responses":{"200":{"description":"The new list of editor IDs (longs).","content":{"application/json":{"schema":{"type":"array","items":{"type":"integer"}}}}},"404":{"description":"Image with specified ID was not found."}}}},"/api/v1/images/folder/{folderId}":{"get":{"operationId":"getByFolderId","parameters":[{"in":"path","name":"folderId","type":"integer","format":"int64","required":true}],"tags":["routes"]}},"/api/v1/comments/imageId/{imageId}":{"get":{"operationId":"getByImageId","tags":["routes"],"parameters":[{"in":"path","name":"imageId","type":"integer","format":"int64","required":true}]},"delete":{"operationId":"deleteByImageId","tags":["routes"],"parameters":[{"in":"path","name":"imageId","type":"integer","format":"int64","required":true}]}},"/":{"get":{"operationId":"index","tags":["routes"]}},"/api/v1/comments/{id}":{"get":{"responses":{"200":{"schema":{"$ref":"#/definitions/models.Comment"}},"404":{"description":"Comment with specified ID was not found"}},"operationId":"getById","parameters":[{"in":"path","name":"id","type":"integer","format":"int64","required":true}],"tags":["routes"]},"delete":{"responses":{"204":{"description":"Comment with specified id has been deleted"},"404":{"description":"Comment with specified ID was not found"}},"operationId":"delete","parameters":[{"in":"path","name":"id","type":"integer","format":"int64","required":true}],"tags":["routes"]}},"/images/{id}":{"get":{"summary":"Returns the image with the given ID.","parameters":[{"in":"path","name":"id","schema":{"type":"integer","minimum":1},"description":"The image ID."}],"responses":{"200":{"description":"A JSON representation of the image with the given ID.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Image"}}}},"404":{"description":"Image with specified ID was not found."}}},"delete":{"summary":"Deletes the image with the given ID.","parameters":[{"in":"path","name":"id","schema":{"type":"integer","minimum":1},"description":"The image ID."}],"responses":{"204":{"description":"Image with the specified ID was deleted."},"404":{"description":"Image with specified ID was not found."}}}},"/images/{id}/folderId":{"patch":{"summary":"Updates the folder ID of the image with the given ID.","parameters":[{"in":"path","name":"id","schema":{"type":"integer","minimum":1},"description":"The image ID."},{"in":"query","name":"folderId","schema":{"type":"integer","description":"The new folder ID."}}],"responses":{"200":{"description":"Newly updated folder ID.","content":{"application/json":{"schema":{"type":"integer"}}}},"404":{"description":"Image with specified ID was not found."}}}},"/images":{"get":{"summary":"Returns a list of all images in the database.","responses":{"200":{"description":"A JSON array of image objects.","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Image"}}}}}}},"post":{"summary":"Creates and returns a new image.","description":"Creates a new image with the given author ID, tags, title and image file. Automatically generates the ID, sets the likes to 0, assigns the author ID to the list of editor IDs and the folder ID to -1.","requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"authorId":{"type":"string"},"tags":{"type":"array","items":{"type":"string"}},"title":{"type":"string"},"image":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"A JSON object of the image.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Image"}}}}}}},"/api/v1/comments/author/{authorId}":{"get":{"responses":{"200":{"schema":{"$ref":"#/definitions/models.Comment"}},"404":{"description":"Comment with specified author ID was not found"}},"operationId":"getByAuthorId","parameters":[{"in":"path","name":"authorId","type":"integer","format":"int64","required":true}],"tags":["routes"]}},"/api/v1/folders":{"post":{"requestBody":{"content":{"application/json":{"schema":{"type":"object","$ref":"#/definitions/dtos.NewFolder"}}}},"responses":{"200":{"schema":{"$ref":"#/definitions/models.Folder"}}},"operationId":"create","tags":["routes"]},"get":{"responses":{"200":{"schema":{"$ref":"#/definitions/models.Folder"}}},"operationId":"getAll","tags":["routes"]}},"/images/{id}/tags":{"patch":{"summary":"Updates the tag list of the image with the given ID.","parameters":[{"in":"path","name":"id","schema":{"type":"integer","minimum":1},"description":"The image ID."},{"in":"query","name":"tags","schema":{"type":"array","items":{"type":"string"},"description":"A list of strings which represent the image tags."}}],"responses":{"200":{"description":"Returns the newly updated tag list.","content":{"application/json":{"schema":{"type":"array","items":{"type":"string"}}}}},"404":{"description":"Image with specified ID was not found."}}}},"/api/v1/folders/{id}/title":{"patch":{"requestBody":{"content":{"application/json":{"schema":{"type":"object","$ref":"#/definitions/models.Folder"}}}},"responses":{"200":{"schema":{"$ref":"#/definitions/models.Folder"}}},"operationId":"updateTitle","parameters":[{"in":"path","name":"id","type":"integer","format":"int64","required":true}],"tags":["routes"]}}},"definitions":{"models.Comment":{"properties":{"imageId":{"format":"int64","type":"integer"},"likeCount":{"format":"int32","type":"integer"},"id":{"format":"int64","type":"integer"},"authorId":{"format":"int64","type":"integer"},"content":{"type":"string"}},"required":["id","authorId","imageId","content","likeCount"]},"models.Folder":{"properties":{"id":{"format":"int64","type":"integer"},"title":{"type":"string"},"authorId":{"format":"int64","type":"integer"}},"required":["id","authorId","title"]}},"info":{"description":"Image sharing webapp API made using the Play Framework, Slick & Scala","title":"Image sharing webapp API","version":"0.0.1"},"tags":[{"name":"routes"}]}